fix pack

+++ src/ITCC.YandexSpeeckKitClient/MessageModels/HttpMode/VariantMessage.cs
@@ -14,7 +14,7 @@ namespace ITCC.YandexSpeeckKitClient.MessageModels.HttpMode
         /// Confidence of hypothesis.
         /// </summary>
         [XmlAttribute(AttributeName = "confidence")]
-        public double Confidence { get; set; }
+        public float Confidence { get; set; }
 
         /// <summary>
         /// The normalized recognized text. In a normalized text, numbers are written as digits, and punctuation and abbreviations are included.
